Initial Octave repository
Two m files are added to create the initial repository.

tren_pulsos.m just create a pulse train.

Filters_design is a script to design a set of Fir and IIR filters and analyse their frequency response. It's intended to give support in filter design and documentation of the library DSP_CRP